Added new bot
Added ADBot.java to iterate over each cell using a DFS approach, to see if the cell is a mine or not, and recognizing the pattern of 1-2 to more easily figure out which cell is a mine. After everything and we're still running into trouble, getSafestCell() will kick in to calculate cell probability and the probability in the surrounding area to see if that cell is safe to mark or not.